rangement des codes pour le wrapping 1-2 cylindres
